+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.3.0
+
+- Fmap.containsKey works faster
+- Fmap.entries works faster
+
 # 0.2.0
 
 - Added .temp() static method